pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["walrus"]

[project]
name = "walrus"
requires-python = ">=3.8"
version = "1.0.0"
readme = "README.md"
dependencies = [
    "einops>=0.8,<0.9",
    "h5py>=3.9.0,<4",
    "hydra-core>=1.3,<2",
    "imageio>=2.25,<3",
    "matplotlib>=3.8,<4",
    "numpy==1.26.4",
    "pandas==2.2",
    "plotly>=5.0,<6",
    "the_well[benchmark] @ git+https://github.com/PolymathicAI/the_well.git@master",
    "timm>=1.0,<2",
    "torch>=2.4, <=2.6",
    "torchinfo>=1.8.0,<2",
    "wandb>=0.17.9"
]


[project.optional-dependencies]
test = [
    "pytest",
    "ruff",
    "isort[colors]>=5.13.2",
    "pytest-order>=1.3.0",
    "mypy>=1.13.0",
    "types-pyyaml>=6",
    "types-redis>=4",
    "types-requests>=2",
    "types-six>=1",
    "types-tabulate>=0.9",
]
external_models = [
    "tokenizers == 0.13.3",
    "transformers == 4.29.2",
    "accelerate == 0.31.0",
    "pyyaml",
]


[tool.isort]
profile = "black"

[tool.mypy]
exclude = '^(walrus/models/external_models/|.*\\.ipynb$)'
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "walrus.optim.distributed_shampoo.*"
ignore_errors = true

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    ]
extend-exclude = [
  "**/*.ipynb",
  "./walrus/models/external_models",
  "walrus/example_run_scripts"
]

[tool.ruff.lint]
fixable = ["ALL"]
